src/third_party/blink/web_tests/fast/forms/file/file-input-key-space-expected.txt

CONSOLE MESSAGE: line 5: FileChooser: opened; mode=kOpen
Ensure pressing Space key with focus on an INPUT element with TYPE=FILE launches a file chooser. Automated test passes if 'FileChooser: opened' was logged.
